Better hash tests
