+++ docs/user/01-sql-statement/06-functions-operators/07-hash-functions/siphash.md
@@ -26,17 +26,17 @@ A UInt64 data type hash value.
 
 ```sql
 mysql> SELECT SIPHASH('1234567890');
-+---------------------+
-| SIPHASH(1234567890) |
-+---------------------+
-| 9027491583908826579 |
-+---------------------+
++-----------------------+
+| SIPHASH('1234567890') |
++-----------------------+
+|  18110648197875983073 |
++-----------------------+
 
 mysql> SELECT SIPHASH(1);
 +---------------------+
 | SIPHASH(1)          |
 +---------------------+
-| 2206609067086327257 |
+| 4952851536318644461 |
 +---------------------+
 
 mysql> SELECT SIPHASH(1.2);
@@ -48,7 +48,7 @@ mysql> SELECT SIPHASH(1.2);
 
 mysql> SELECT SIPHASH(number) FROM numbers(2);
 +----------------------+
-| siphash(number)      |
+| SIPHASH(number)      |
 +----------------------+
 | 13646096770106105413 |
 |  2206609067086327257 |
